+++ R/protected_api-FutureCondition-class.R
@@ -223,6 +223,44 @@ GlobalEnvFutureError <- function(...) {
 
 #' @rdname FutureCondition
 #' @export
+ConnectionsMisuseFutureCondition <- function(message = NULL, call = NULL, differences = NULL, uuid = future[["uuid"]], future = NULL) {
+  if (is.null(message)) {
+    label <- sQuoteLabel(future[["label"]])
+    if (is.null(differences)) {
+      info <- "added, removed, or modified connections"
+    } else {
+      info <- sprintf("added (n=%d), removed (n=%d), or replaced (n=%d) connections", differences[["added"]], differences[["removed"]], differences[["replaced"]])
+    }
+    message <- sprintf("Future (%s) %s. A future expression must close any opened connections and must not close connections it did not open", label, info)
+  }
+  cond <- FutureCondition(message = message, call = call, uuid = uuid, future = future)
+  cond[["differences"]] <- differences
+  class <- c("ConnectionsMisuseFutureCondition", "MisuseFutureCondition", class(cond))
+  class(cond) <- class[!duplicated(class, fromLast = TRUE)]
+  cond
+}
+
+#' @rdname FutureCondition
+#' @export
+ConnectionsMisuseFutureWarning <- function(...) {
+  cond <- ConnectionsMisuseFutureCondition(...)
+  class <- c("ConnectionsMisuseFutureWarning", "MisuseFutureWarning", "FutureWarning", "warning", class(cond))
+  class(cond) <- class[!duplicated(class, fromLast = TRUE)]
+  cond
+}
+
+#' @rdname FutureCondition
+#' @export
+ConnectionsMisuseFutureError <- function(...) {
+  cond <- ConnectionsMisuseFutureCondition(...)
+  class <- c("ConnectionsMisuseFutureError", "MisuseFutureError", "FutureError", "error", class(cond))
+  class(cond) <- class[!duplicated(class, fromLast = TRUE)]
+  cond
+}
+
+
+#' @rdname FutureCondition
+#' @export
 FutureInterruptError <- function(..., future = NULL) {
   cond <- FutureError(..., future = future)
   class <- c("FutureInterruptError", "FutureError", class(cond))