+++ R/utils-options.R
@@ -61,6 +61,8 @@
 #'
 #'  \item{\option{future.rng.onMisuse}: (_beta feature - may change_)}{(character string) If random numbers are used in futures, then parallel (L'Ecuyer-CMRG) RNG should be used in order to get statistical sound RNGs. The defaults in the future framework assume that _no_ random number generation (RNG) is taken place in the future expression because L'Ecuyer-CMRG RNGs come with an unnecessary overhead if not needed.  To protect against mistakes, the future framework attempts to detect when random numbers are used despite L'Ecuyer-CMRG RNGs are not in place.  If this is detected, and `future.rng.onMisuse = "error"`, then an informative error message is produced.  If `"warning"`, then a warning message is produced.  If `"ignore"`, no check is performed. (Default: `"warning"`)}
 #'
+#'  \item{\option{future.connections.onMisuse}: (_beta feature - may change_)}{(character string) A future must close any connections it opens and must not close connections it did not open. If such misuse is detected and this option is set to `"error"`, `value()` will produce an error with details. If it is set to `"warning"`, a warning is produced. If `"ignore"`, no check is performed. (Default: `"warning"`)}
+#'
 #'  \item{\option{future.globalenv.onMisuse}: (_beta feature - may change_)}{(character string) Assigning variables to the global environment for the purpose of using the variable at a later time makes no sense with futures, because the next future may be evaluated in different R process.  To protect against mistakes, the future framework attempts to detect when variables are added to the global environment.  If this is detected, and `future.globalenv.onMisuse = "error"`, then an informative error message is produced.  If `"warning"`, then a warning message is produced.  If `"ignore"`, no check is performed. (Default: `"ignore"`)}
 #'
 #'  \item{\option{future.onFutureCondition.keepFuture}:}{(logical) If `TRUE`, a `FutureCondition` keeps a copy of the `Future` object that triggered the condition. If `FALSE`, it is dropped. (Default: `TRUE`)}
@@ -161,6 +163,7 @@
 #' future.plan
 #' future.onFutureCondition.keepFuture
 #' future.resolve.recursive
+#' future.connections.onMisuse
 #' future.globalenv.onMisuse
 #' future.rng.onMisuse
 #' future.wait.alpha
@@ -183,6 +186,7 @@
 #' R_FUTURE_PLAN
 #' R_FUTURE_ONFUTURECONDITION_KEEPFUTURE
 #' R_FUTURE_RESOLVE_RECURSIVE
+#' R_FUTURE_CONNECTIONS_ONMISUSE
 #' R_FUTURE_GLOBALENV_ONMISUSE
 #' R_FUTURE_RNG_ONMISUSE
 #' R_FUTURE_WAIT_ALPHA
@@ -367,4 +371,7 @@ update_package_options <- function(debug = FALSE) {
   ## future 1.34.0:
   update_package_option("future.globals.objectSize.method", mode = "character", debug = debug)
   update_package_option("future.plan.cleanup.legacy", mode = "logical", debug = debug)
+
+  ## future (> 1.34.0-9000):
+  update_package_option("future.connections.onMisuse", mode = "character", debug = debug)
 }